PythonScripts/requirements.txt
gpiozero
matplotlib
datetime
RPI.GPIO